worked on the next part of project 07 and the Challenge for the section
This section  I was learning alternatives for prop drilling and how to make props more reusable. Then we also went ahead and made the StarRating component using what I learned and tried to make it in a way it would be more widly reusable than if I had made it just for the one usecase of project 7. I did the challenge collapse text that was found on code sandbox.